CfgPatches für Eric aufgeräumt; Version hochgesetzt für nächsten Release; weitere Konfigs in Main abgespalten
