--- TB_MAINCORE.VR/stuff/README_Commands.txt
@@ -1,1 +1,0 @@
-https://github.com/shukari/TBMod/wiki/SkripteSammlung
