+++ addons/TBMod_arsenal/config.cpp
@@ -4,7 +4,6 @@ class CfgPatches
     {
         name = "TBMod Arsenal";
         
-        weapons[] = {};
         units[] = {
             "TB_arsenal_base",
             "TB_arsenal_usa",
@@ -14,7 +13,6 @@ class CfgPatches
             "TB_arsenal_themen",
             "TB_arsenal_predefined_custom"
         };
-        ammo[] = {};
         
         requiredAddons[] = {"TBMod_main"};
         addonRootClass = "TBMod_main";
