+++ addons/TBMod_main/config.cpp
@@ -1,3 +1,7 @@
+/*
+    Part of the TBMod ( https://github.com/shukari/TBMod )
+    Developed by http://tacticalbacon.de
+*/
 class CfgPatches
 {
     class TBMod_main
@@ -7,13 +11,12 @@ class CfgPatches
         author = "shukari";
         authorUrl = "tacticalbacon.de";
 
-        version = 1.1.2;
-        versionStr = "1.1.2";
-        versionAr[] = {1,1,2};
+        version = 1.1.3;
+        versionStr = "1.1.3";
+        versionAr[] = {1,1,3};
         versionDesc = "TBMod";
         versionAct = "";
 
-        weapons[] = {};
         units[] = {
             "TB_flag_base",
             "TB_flag_rufnamen",
@@ -77,39 +80,8 @@ class CfgPatches
 #include "configs\CfgUnitInsignia.hpp"
 #include "configs\CfgVehicles.hpp"
 #include "configs\ExtendedEventHandlers.hpp"
+#include "configs\CBA_Stuff.hpp"
+#include "configs\CfgMisc.hpp"
 
 // Other config stuff
 #include "gcam\gcam_component.hpp"
-
-// CBA STUFF
-cba_settings_whitelist[] = {"76561198029318101", "admin"};
-enableTargetDebug = 1;
-enableDebugConsole[] = {"76561198029318101"};
-
-class cfgScriptPaths
-{
-    TB_Main = "\TBMod_main\functions\";
-    TB_replacement = "\TBMod_main\replacement\";
-};
-
-class RscDisplayGarage;
-class TB_RscDisplayGarage: RscDisplayGarage
-{
-    enableSimulation = 1;
-    icon = "\A3\Ui_f\data\Logos\a_64_ca.paa";
-    idd = -1;
-    logo = "\A3\Ui_f\data\Logos\arsenal_1024_ca.paa";
-    onLoad = "['onLoad',_this,'TB_RscDisplayGarage','GUI'] call  (uinamespace getvariable 'BIS_fnc_initDisplay')";
-    onUnload = "['onUnload',_this,'TB_RscDisplayGarage','GUI'] call  (uinamespace getvariable 'BIS_fnc_initDisplay')";
-    scriptName = "TB_RscDisplayGarage";
-    scriptPath = "TB_Main";
-};
-
-class RscDisplayCurator
-{
-    idd = 312;
-    onLoad = "[_this select 0] call Achilles_fnc_onDisplayCuratorLoad;";
-    onUnload = "[_this select 0] call Achilles_fnc_onDisplayCuratorUnload;";
-    scriptName = "RscDisplayCurator";
-    scriptPath = "TB_replacement";
-};
