+++ addons/TBMod_main/configs/CBA_Stuff.hpp
@@ -1,0 +1,7 @@
+/*
+    Part of the TBMod ( https://github.com/shukari/TBMod )
+    Developed by http://tacticalbacon.de
+*/
+cba_settings_whitelist[] = {"76561198029318101", "admin"};
+enableTargetDebug = 1;
+enableDebugConsole[] = {"76561198029318101"};
