+++ addons/TBMod_nachschub/config.cpp
@@ -4,10 +4,6 @@ class CfgPatches
     {
         name = "TBMod Nachschub";
         
-        weapons[] = {};
-        units[] = {};
-        ammo[] = {};
-        
         requiredAddons[] = {"TBMod_main"};
         addonRootClass = "TBMod_main";
     };
