+++ addons/TBMod_unsung/config.cpp
@@ -4,10 +4,6 @@ class CfgPatches
     {
         name = "TBMod Unsung";
         
-        weapons[] = {};
-        units[] = {};
-        ammo[] = {};
-        
         requiredAddons[] = {
             "TBMod_nachschub",
             "uns_main"
